+++ src/Sulu/Bundle/MediaBundle/Content/Types/MediaSelectionContentType.php
@@ -116,22 +116,7 @@ class MediaSelectionContentType extends ComplexContentType
      */
     private function setData($data, PropertyInterface $property, $languageCode)
     {
-        if ($data instanceof MediaSelectionContainer) {
-            $container = $data;
-        } else {
-            $params = $this->getParams($property->getParams());
-            $types = $params['types'];
-            $container = new MediaSelectionContainer(
-                isset($data['config']) ?  $data['config'] : array(),
-                isset($data['displayOption']) ? $data['displayOption'] : '',
-                isset($data['ids']) ? $data['ids'] : array(),
-                $languageCode,
-                $types,
-                $this->mediaManager
-            );
-        }
-
-        $property->setValue($container);
+        $property->setValue($data);
     }
 
     /**
@@ -169,7 +154,9 @@ class MediaSelectionContentType extends ComplexContentType
         $languageCode,
         $segmentKey
     ) {
-        $node->getProperty($property->getName())->remove();
+        if ($node->hasProperty($property->getName())) {
+            $node->getProperty($property->getName())->remove();
+        }
     }
 
     /**
@@ -185,7 +172,21 @@ class MediaSelectionContentType extends ComplexContentType
      */
     public function getContentData(PropertyInterface $property)
     {
-        return $property->getValue()->getData();
+        $data = $property->getValue();
+
+        $params = $this->getParams($property->getParams());
+        $types = $params['types'];
+
+        $container = new MediaSelectionContainer(
+            isset($data['config']) ?  $data['config'] : array(),
+            isset($data['displayOption']) ? $data['displayOption'] : '',
+            isset($data['ids']) ? $data['ids'] : array(),
+            $property->getStructure()->getLanguageCode(),
+            $types,
+            $this->mediaManager
+        );
+
+        return $container->getData();
     }
 
     /**
@@ -193,6 +194,6 @@ class MediaSelectionContentType extends ComplexContentType
      */
     public function getViewData(PropertyInterface $property)
     {
-        return $property->getValue()->toArray();
+        return $property->getValue();
     }
 }