+++ src/Sulu/Bundle/SnippetBundle/Controller/SnippetController.php
@@ -10,14 +10,18 @@
 
 namespace Sulu\Bundle\SnippetBundle\Controller;
 
+use PHPCR\NodeInterface;
 use Sulu\Component\Content\Mapper\ContentMapper;
 use Sulu\Component\Content\StructureInterface;
 use Sulu\Component\Content\StructureManager;
+use Sulu\Component\Rest\ListBuilder\ListRepresentation;
+use Sulu\Component\Rest\ListBuilder\ListRestHelper;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Sulu\Bundle\SnippetBundle\Snippet\SnippetRepository;
 use FOS\RestBundle\View\ViewHandler;
 use FOS\RestBundle\View\View;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Core\SecurityContext;
 use Sulu\Component\Content\Mapper\ContentMapperRequest;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
@@ -58,6 +62,9 @@ class SnippetController
      */
     protected $urlGenerator;
 
+    /**
+     * @var string
+     */
     protected $languageCode;
 
     /**
@@ -88,22 +95,34 @@ class SnippetController
     {
         $this->initEnv($request);
 
+        $listRestHelper = new ListRestHelper($request);
+
         // if the type parameter is falsy, assign NULL to $type
-        $type = $request->query->get('type', null) ? : null;
+        $type = $request->query->get('type', null) ?: null;
+
         $uuidsString = $request->get('ids');
 
         if ($uuidsString) {
             $uuids = explode(',', $uuidsString);
             $snippets = $this->snippetRepository->getSnippetsByUuids($uuids, $this->languageCode);
+            $total = count($snippets);
         } else {
             $snippets = $this->snippetRepository->getSnippets(
                 $this->languageCode,
                 $type,
-                null,
-                null,
-                $request->get('search'),
-                $request->get('sortBy'),
-                $request->get('sortOrder')
+                $listRestHelper->getOffset(),
+                $listRestHelper->getLimit(),
+                $listRestHelper->getSearchPattern(),
+                $listRestHelper->getSortColumn(),
+                $listRestHelper->getSortOrder()
+            );
+
+            $total = $this->snippetRepository->getSnippetsAmount(
+                $this->languageCode,
+                $type,
+                $listRestHelper->getSearchPattern(),
+                $listRestHelper->getSortColumn(),
+                $listRestHelper->getSortOrder()
             );
         }
 
@@ -113,11 +132,17 @@ class SnippetController
             $data[] = $snippet->toArray();
         }
 
-        $data = $this->decorateList($data, $this->languageCode);
+        $data = new ListRepresentation(
+            $this->decorateSnippets($data, $this->languageCode),
+            'snippets',
+            'get_snippets',
+            $request->query->all(),
+            $listRestHelper->getPage(),
+            $listRestHelper->getLimit(),
+            $total
+        );
 
-        $view = View::create($data);
-
-        return $this->viewHandler->handle($view);
+        return $this->viewHandler->handle(View::create($data));
     }
 
     /**
@@ -281,20 +306,20 @@ class SnippetController
             )
         );
     }
-
+    
     /**
-     * Returns user
-     */
-    private function getUser()
-    {
-        $token = $this->securityContext->getToken();
+    * Returns user
+    */
+   private function getUser()
+   {
+       $token = $this->securityContext->getToken();
 
-        if (null === $token) {
-            throw new \InvalidArgumentException('No user is set');
-        }
-
-        return $token->getUser();
-    }
+       if (null === $token) {
+           throw new \InvalidArgumentException('No user is set');
+       }
+       
+       return $token->getUser();
+   }
 
     /**
      * Initiates the environment
@@ -328,42 +353,6 @@ class SnippetController
     }
 
     /**
-     * Decorate the list for HATEOAS
-     *
-     * TODO: Use the HateoasBundle / JMSSerializer to do this.
-     */
-    private function decorateList(array $data, $locale)
-    {
-        return array(
-            'page' => 1,
-            'limit' => PHP_INT_MAX,
-            'pages' => 1,
-            'total' => count($data),
-            '_links' => array(
-                'self' => array(
-                    'href' => $this->urlGenerator->generate('get_snippets', array('language' => $locale)),
-                ),
-                'find' => array(
-                    'href' => $this->urlGenerator->generate(
-                            'get_snippets',
-                            array('language' => $locale, 'search' => '{searchString}')
-                        ),
-                ),
-                'sortable' => array(
-                    'href' => $this->urlGenerator->generate(
-                            'get_snippets',
-                            array('language' => $locale, 'sortBy' => '{sortBy}', 'sortOrder' => '{sortOrder}')
-                        ),
-                )
-            ),
-            '_embedded' => array(
-                'snippets' => $this->decorateSnippets($data, $locale)
-            ),
-        );
-
-    }
-
-    /**
      * Decorate snippets for HATEOAS
      */
     private function decorateSnippets(array $snippets, $locale)
@@ -386,18 +375,18 @@ class SnippetController
             array(
                 '_links' => array(
                     'self' => $this->urlGenerator->generate(
-                            'get_snippet',
-                            array('uuid' => $snippet['id'], 'language' => $locale)
-                        ),
+                        'get_snippet',
+                        array('uuid' => $snippet['id'], 'language' => $locale)
+                    ),
                     'delete' => $this->urlGenerator->generate(
-                            'delete_snippet',
-                            array('uuid' => $snippet['id'], 'language' => $locale)
-                        ),
+                        'delete_snippet',
+                        array('uuid' => $snippet['id'], 'language' => $locale)
+                    ),
                     'new' => $this->urlGenerator->generate('post_snippet', array('language' => $locale)),
                     'update' => $this->urlGenerator->generate(
-                            'put_snippet',
-                            array('uuid' => $snippet['id'], 'language' => $locale)
-                        ),
+                        'put_snippet',
+                        array('uuid' => $snippet['id'], 'language' => $locale)
+                    ),
                 ),
             )
         );
@@ -423,7 +412,12 @@ class SnippetController
 
         foreach ($references as $reference) {
             if ($reference->getParent()->isNodeType('sulu:page')) {
-                $content = $this->contentMapper->load($reference->getParent()->getIdentifier(), $webspace, $this->languageCode, true);
+                $content = $this->contentMapper->load(
+                    $reference->getParent()->getIdentifier(),
+                    $webspace,
+                    $this->languageCode,
+                    true
+                );
                 $data['structures'][] = $content->toArray();
             } else {
                 $data['other'] = $reference->getPath();