+++ tests/app/AppKernel.php
@@ -6,4 +6,11 @@ use Sulu\Bundle\TestBundle\Kernel\SuluTestKernel;
 
 class AppKernel extends SuluTestKernel
 {
+
+    public function registerBundles()
+    {
+        $bundles = parent::registerBundles();
+        $bundles[] = new \Sulu\Bundle\SnippetBundle\SuluSnippetBundle();
+        return $bundles;
+    }
 }